Added experimental renderer interface
Before this change, all rendering logic (for OpenGL and the various tracers) was in SLSceneView, which made the class quite big. Now every rendering backend has its own implementation of the SLRenderer interface. SLSceneView then selects the correct renderer based on the current SLRenderType and calls its virtual render function every frame. The new interface should make it easier to add a new renderer (for example for the Vulkan API or for a VR headset) because one only needs to implement the SLRenderer interface and add a new value to the SLRenderType enum. It should now also be easier to change an existing rendering backend, because all the related code is grouped in the relevant subclass.